🔖 Prepare for release v2.0.2
