+++ .envrc
@@ -27,6 +27,7 @@ export MAX_ROWS=1 # Setting for simplecov-console gem for tty output, limits to 
 
 # Internal Debugging Controls
 export DEBUG=false # do not allow byebug statements (override in .env.local)
+export REQUIRE_BENCH=false # set to true in .env.local to turn on require_bench
 
 # .env would override anything in this file, if `dotenv` is uncommented below.
 # .env is a DOCKER standard, and if we use it, it would be in deployed, or DOCKER, environments,
