+++ .simplecov
@@ -1,3 +1,5 @@
 require "kettle/soup/cover/config"
 
-SimpleCov.start
+SimpleCov.start do
+  add_filter "test/**/*"
+end
