+++ CHANGELOG.md
@@ -10,9 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Removed
 
-## 2.0.2 - 2024-09-20
+## 2.0.2 - 2024-09-25
+- COVERAGE:  97.72% -- 214/219 lines in 4 files
+- BRANCH COVERAGE:  86.00% -- 43/50 branches in 4 files
+- 39.13% documented
 ### Added
 - Improved integration with direnv for development
+- Better test coverage
+- More tests
+- rots v1.0.0 for tests
+### Changed
+- Upgraded to ruby-openid2 v3.1.0
+- Moved to oauth-xx organization
 
 ## 2.0.1 - 2024-09-05
 ### Added
