+++ lib/rack/openid/version.rb
@@ -1,7 +1,7 @@
 module Rack
   class OpenID
     module Version
-      VERSION = "2.0.1"
+      VERSION = "2.0.2"
     end
   end
 end
